requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.6.1
fastapi==0.95.1
h11==0.14.0
idna==3.4
pip==23.0.1
pydantic==1.10.7
pymongo==4.7.1
python-dotenv==1.0.1
setuptools==67.6.1
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
wheel==0.40.0